refactor: use strata_asm_proto_upgrade_txs
