+++ crates/asm/subprotocols/upgrade/src/multisig/mod.rs
@@ -1,4 +1,1 @@
 pub mod authority;
-pub mod config;
-pub mod msg;
-pub mod vote;
